requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0
pytest>=7.0.0
